frontend/requirements.docker.txt
# Docker-specific requirements file
# This file is used only for Docker builds where the shared library is at ./shared

# First install the shared library from the correct Docker context path
-e file:./shared

# External dependencies (copied from requirements.txt to avoid the ../shared path)
streamlit>=1.28.1,<1.29.0
httpx>=0.28.1
pydantic-settings>=2.0.3,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.1,<0.24.0 